move sourcemap tests to separate module, test out different configurations
