+++ packages/replace/test/snapshots/sourcemaps.js.md
@@ -1,0 +1,185 @@
+# Snapshot report for `test/sourcemaps.js`
+
+The actual snapshot is saved in `sourcemaps.js.snap`.
+
+Generated by [AVA](https://ava.li).
+
+## generates sourcemaps when sourcemap is used as an output argument
+
+> Snapshot 1
+
+    {
+      character: 12,
+      column: 12,
+      line: 1,
+    }
+
+> Snapshot 2
+
+    {
+      column: 15,
+      line: 1,
+      name: null,
+      source: 'other.js',
+    }
+
+> Snapshot 3
+
+    {
+      character: 17,
+      column: 0,
+      line: 3,
+    }
+
+> Snapshot 4
+
+    {
+      column: 0,
+      line: 2,
+      name: null,
+      source: 'main.js',
+    }
+
+## generates sourcemaps if enabled in plugin
+
+> Snapshot 1
+
+    {
+      character: 12,
+      column: 12,
+      line: 1,
+    }
+
+> Snapshot 2
+
+    {
+      column: 15,
+      line: 1,
+      name: null,
+      source: 'other.js',
+    }
+
+> Snapshot 3
+
+    {
+      character: 17,
+      column: 0,
+      line: 3,
+    }
+
+> Snapshot 4
+
+    {
+      column: 0,
+      line: 2,
+      name: null,
+      source: 'main.js',
+    }
+
+## generates sourcemaps if enabled in plugin (camelcase)
+
+> Snapshot 1
+
+    {
+      character: 12,
+      column: 12,
+      line: 1,
+    }
+
+> Snapshot 2
+
+    {
+      column: 15,
+      line: 1,
+      name: null,
+      source: 'other.js',
+    }
+
+> Snapshot 3
+
+    {
+      character: 17,
+      column: 0,
+      line: 3,
+    }
+
+> Snapshot 4
+
+    {
+      column: 0,
+      line: 2,
+      name: null,
+      source: 'main.js',
+    }
+
+## generates sourcemaps by dfault
+
+> Snapshot 1
+
+    {
+      character: 12,
+      column: 12,
+      line: 1,
+    }
+
+> Snapshot 2
+
+    {
+      column: 15,
+      line: 1,
+      name: null,
+      source: 'other.js',
+    }
+
+> Snapshot 3
+
+    {
+      character: 17,
+      column: 0,
+      line: 3,
+    }
+
+> Snapshot 4
+
+    {
+      column: 0,
+      line: 2,
+      name: null,
+      source: 'main.js',
+    }
+
+## generates sourcemaps by default
+
+> Snapshot 1
+
+    {
+      character: 12,
+      column: 12,
+      line: 1,
+    }
+
+> Snapshot 2
+
+    {
+      column: 15,
+      line: 1,
+      name: null,
+      source: 'other.js',
+    }
+
+> Snapshot 3
+
+    {
+      character: 17,
+      column: 0,
+      line: 3,
+    }
+
+> Snapshot 4
+
+    {
+      column: 0,
+      line: 2,
+      name: null,
+      source: 'main.js',
+    }
